feat : refresh 시 토큰이 부적절하다면 401 에러를 반환합니다.

+++ friends_server/src/main/kotlin/com/friends/security/service/AuthServices.kt
@@ -1,9 +1,11 @@
 package com.friends.security.service
 
+import com.friends.common.exception.ErrorCode
 import com.friends.jwt.AuthJwtRepository
 import com.friends.jwt.JwtService
 import com.friends.member.entity.Member
 import com.friends.member.repository.MemberRepository
+import com.friends.security.securityException.InvalidJwtException
 import com.friends.security.userDetails.CustomUserDetails
 import org.springframework.security.authentication.AuthenticationManager
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
@@ -31,6 +33,10 @@ class AuthService(
     }
 
     fun refresh(refreshToken: String): AtRtDto {
+        if (!jwtService.validateRefreshToken(refreshToken)) {
+            throw InvalidJwtException(ErrorCode.INVALID_TOKEN)
+        }
+
         // refresh 될 때 기존의 access token 과 refresh token 을 삭제합니다.
         authJwtRepository.deleteRefreshToken(refreshToken)
         authJwtRepository.getAccessToken(refreshToken)?.apply {
